Command line parameters for features; moving blocks.

+++ ge.demo/src/ge/demo/actor/Player.java
@@ -18,7 +18,7 @@ public class Player extends Actor
 		
 	}
 	
-	public void act(int delta)
+	public void act(long delta)
 	{
 		
 	}
