+++ ge.demo/src/ge/demo/actor/Purploid.java
@@ -15,7 +15,7 @@ public class Purploid extends Actor
 		
 	}
 	
-	public void act(int delta)
+	public void act(long delta)
 	{
 		
 	}
